Refactor to move create_bigquery_table to listenbrainz.bigquery
This will allow easier usage in bigquery_writer.
Also, use globally defined constants for table names like
'before_2002'.
